Refactor code to align with code quality standards
Code in the repository can always be looked for improvement. Aligning to
standards such as LoC for method lengths, SLAP and not deeply nesting is
important to ensure high quality of code.

By properly following good coding quality practices, it ensures
readability of code and maintains standards for future features.

Let's refactor the parse method in the InputParser class to align to
these coding standards by reducing the LoC.

In order to reduce the LoC in the method:
* Common checks such as checking the number of arguments in input parts
  will be extracted to its individual method, validateArguments.
* For each case in the switch, extract the check of number of arguments
  and returning of the Command to its respective method.

Through extracting commonly used code, we heavily reduce the number of
lines required in the original parse method and align with coding
quality standards.